[NAYB-96] refactor : Item 패키지 구조 변경
패키지 컨벤션에 맞도록 Item 엔티티 위치 변경
