+++ src/main/java/com/prgrms/nabmart/domain/cart/CartItem.java
@@ -3,7 +3,7 @@ package com.prgrms.nabmart.domain.cart;
 import static java.util.Objects.isNull;
 
 import com.prgrms.nabmart.domain.cart.exception.InvalidCartItemException;
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import com.prgrms.nabmart.global.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
