+++ src/main/java/com/prgrms/nabmart/domain/category/MainCategory.java
@@ -1,6 +1,6 @@
 package com.prgrms.nabmart.domain.category;
 
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
