+++ src/main/java/com/prgrms/nabmart/domain/event/repository/EventItemRepository.java
@@ -2,7 +2,7 @@ package com.prgrms.nabmart.domain.event.repository;
 
 import com.prgrms.nabmart.domain.event.domain.Event;
 import com.prgrms.nabmart.domain.event.domain.EventItem;
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
