+++ src/main/java/com/prgrms/nabmart/domain/order/OrderItem.java
@@ -1,6 +1,6 @@
 package com.prgrms.nabmart.domain.order;
 
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import com.prgrms.nabmart.global.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
