+++ src/main/java/com/prgrms/nabmart/domain/review/Review.java
@@ -1,6 +1,6 @@
 package com.prgrms.nabmart.domain.review;
 
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import com.prgrms.nabmart.domain.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
