+++ src/test/java/com/prgrms/nabmart/domain/event/service/EventServiceTest.java
@@ -12,8 +12,8 @@ import com.prgrms.nabmart.domain.event.domain.EventItem;
 import com.prgrms.nabmart.domain.event.repository.EventRepository;
 import com.prgrms.nabmart.domain.event.service.request.FindEventDetailCommand;
 import com.prgrms.nabmart.domain.event.service.request.RegisterEventCommand;
-import com.prgrms.nabmart.domain.item.domain.Item;
 import com.prgrms.nabmart.domain.event.support.EventFixture;
+import com.prgrms.nabmart.domain.item.Item;
 import com.prgrms.nabmart.global.fixture.ItemFixture;
 import java.util.Arrays;
 import java.util.List;
